removing further circular dependence between scripts
