+++ expressyeaself/build_promoter.py
@@ -5,9 +5,11 @@ For example: ATGCATGC inserted into AAAANNNNNNNNTTTT would give
 AAAAATGCATGCTTTT.
 """
 import expressyeaself.organize_data as organize
-from expressyeaself.utilities import check_valid_line as check_valid_line
-from expressyeaself.utilities import get_time_stamp as get_time_stamp
-from expressyeaself.utilities import smart_open as smart_open
+import expressyeaself.utilities as utilities  # noqa: F401
+from utilities import check_valid_line as check_valid_line
+from utilities import get_time_stamp as get_time_stamp
+from utilities import separate_seq_and_el_data as separate_seq_and_el_data
+from utilities import smart_open as smart_open
 import os
 # import pandas as pd
 # import xlrd
@@ -115,7 +117,7 @@ def remove_flanks_from_all_seqs(input_seqs, scaffold_type='pTpA'):
         line = check_valid_line(line)
         if line == 'skip_line':
             continue
-        seq, exp_level = organize.separate_seq_and_el_data(line)
+        seq, exp_level = separate_seq_and_el_data(line)
         deflanked_seq = remove_flanks_from_seq(seq, scaffold_type)
         outfile.write(deflanked_seq + '\t' + str(exp_level) + '\n')
     # Close the input and output files.
@@ -211,7 +213,7 @@ def insert_all_seq_into_one_scaffold(input_seqs, scaffold_type='pTpA'):
         line = check_valid_line(line)
         if line == 'skip_line':
             continue
-        seq, exp_level = organize.separate_seq_and_el_data(line)
+        seq, exp_level = separate_seq_and_el_data(line)
         complete_seq = insert_seq_into_scaffold(seq, scaffold)
         outfile.write(complete_seq + '\t' + str(exp_level) + '\n')
     # Close the input, output, and scaffold files.
@@ -270,7 +272,7 @@ def pad_sequences(input_seqs, pad_front=False, extra_padding=0):
             line = check_valid_line(line)
             if line == 'skip_line':
                 continue
-            seq, exp_level = organize.separate_seq_and_el_data(line)
+            seq, exp_level = separate_seq_and_el_data(line)
             difference = pad_length - len(seq)
             if difference == 0:
                 padded_seq = seq
