+++ expressyeaself/encode_sequences.py
@@ -5,7 +5,9 @@ neural network that will receive the encoded sequence.
 """
 import numpy as np
 import expressyeaself.organize_data as organize
-from expressyeaself.utilities import check_valid_line as check_valid_line
+import expressyeaself.utilities as utilities  # noqa: F401
+from utilities import check_valid_line as check_valid_line
+from utilities import separate_seq_and_el_data as separate_seq_and_el_data
 from expressyeaself.utilities import smart_open as smart_open
 
 BASES = ['A', 'T', 'G', 'C']
@@ -106,7 +108,7 @@ def encode_sequences_with_method(input_seqs, method='One-Hot',
         line = check_valid_line(line)
         if line == 'skip_line':
             continue  # skip line if not a valid line
-        seq, exp_level = organize.separate_seq_and_el_data(line)
+        seq, exp_level = separate_seq_and_el_data(line)
         # Encode with One-Hot method
         if method == 'One-Hot':
             try:
